feature integrated with sync upstream codebase and ready for PR
