--- api/.env.template
@@ -1,19 +1,0 @@
-# OpenAI/Openrouter Keys
-OPENAI_API_KEY=
-MODEL=
-
-# NextJS & fastAPI
-URL=http://localhost:3000
-HONCHO_URL=https://demo.honcho.dev
-HONCHO_APP_NAME=
-
-# Sentry
-SENTRY_DSN_API=
-SENTRY_ENVIORNMENT=
-
-# Supabase
-SUPABASE_URL=<Supabase Project URL>
-SUPABASE_KEY=<Supabase Service Key
-
-# Stripe
-STRIPE_ENABLED=
