+++ www/.env.template
@@ -1,18 +1,25 @@
 # Core
 NEXT_PUBLIC_SITE_URL=
 NEXT_PUBLIC_API_URL=
+
 # Supabase
 NEXT_PUBLIC_SUPABASE_URL=
 NEXT_PUBLIC_SUPABASE_ANON_KEY=
-SUPABASE_SERVICE_ROLE_KEY
-# Sentry
+SUPABASE_SERVICE_ROLE_KEY=
+
+# JINA AI READER
+NEXT_PUBLIC_JINA_API=
+
+# Sentry (optional)
 NEXT_PUBLIC_SENTRY_DSN=
 SENTRY_ENVIRONMENT=
 SENTRY_RELEASE=
-# Posthog
+
+# Posthog (optional)
 NEXT_PUBLIC_POSTHOG_KEY=
 NEXT_PUBLIC_POSTHOG_HOST=
-# Stripe
+
+# Stripe (optional)
 NEXT_PUBLIC_STRIPE_ENABLED=false
 STRIPE_SECRET_KEY=
 NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=
